64-bit fixes

+++ System.Data.SQLite/SQLiteStatement.cs
@@ -27,7 +27,7 @@ namespace System.Data.SQLite
     /// <summary>
     /// The actual statement pointer
     /// </summary>
-    internal int               _sqlite_stmt;
+    internal IntPtr            _sqlite_stmt;
     /// <summary>
     /// An index from which unnamed parameters begin
     /// </summary>
@@ -52,7 +52,7 @@ namespace System.Data.SQLite
     /// <param name="stmt">The statement</param>
     /// <param name="strCommand">The command text for this statement</param>
     /// <param name="previous">The previous command in a multi-statement command</param>
-    internal SQLiteStatement(SQLiteBase sqlbase, int stmt, string strCommand, SQLiteStatement previous)
+    internal SQLiteStatement(SQLiteBase sqlbase, IntPtr stmt, string strCommand, SQLiteStatement previous)
     {
       _sql     = sqlbase;
       _sqlite_stmt = stmt;
